converter, compute: fix static analysis style findings
Several linters flag style issues that must be resolved before the
package can be added to the linter's scope.

- Lowercase error strings and remove trailing punctuation (ST1005)
- Remove an unnecessary uint() cast (unconvert)
- Simplify var-with-type declarations to short form (ST1023)
- Replace len(s) > 0 with s != "" (S1009)

+++ pkg/virt-launcher/virtwrap/converter/compute/controllers.go
@@ -68,7 +68,7 @@ func (c ControllersDomainConfigurator) Configure(vmi *v1.VirtualMachineInstance,
 	}
 
 	if requiresSCSIController(vmi) {
-		scsiControllerDriver := assignSCSIControllerIOThread(vmi, uint(c.autoThreads), controllerDriver.DeepCopy())
+		scsiControllerDriver := assignSCSIControllerIOThread(vmi, c.autoThreads, controllerDriver.DeepCopy())
 		domain.Spec.Devices.Controllers = append(domain.Spec.Devices.Controllers, newSCSIController(c.scsiModel, scsiControllerDriver))
 	}
